Closes issue #695

+++ xga/samples/base.py
@@ -284,9 +284,9 @@ class BaseSample:
         """
         Simple function to show basic information about the sample.
         """
-        # Finding the number of sources in the sample that have been detected in AT LEAST on ObsID
+        # Finding the number of sources in the sample that have been detected in AT LEAST one ObsID
         num_det = sum(np.array([sum(s.detected.values()) for s in self._sources.values()]) >= 1)
-        perc_det = round(num_det / len(self._sources), 2) * 100
+        perc_det = int(round(num_det / len(self._sources), 2) * 100)
         print("\n-----------------------------------------------------")
         print("Number of Sources - {}".format(len(self)))
         print("Redshift Information - {}".format(self.redshifts[0] is not None))
